requirements.txt
# Discord Bot - Production Dependencies
# Updated: 2026-04-12 (security audit pass)
# Python 3.14+

# ==================== Core ====================
discord.py==2.7.1             # Discord API wrapper; voice crypto is pinned separately below
python-dotenv==1.2.1          # Environment variable management (latest)
aiohttp==3.13.4               # Async HTTP client (security fixes)
psutil==7.2.2                 # Process and system utilities
certifi>=2026.1.4             # SSL certificates (security updates)
cryptography>=46.0.7          # Transitive dep — pinned for CVE-2026-34073, CVE-2026-39892
requests>=2.33.0              # Transitive dep — pinned for CVE-2026-25645
tzdata>=2024.1                # IANA timezone data (required on Windows)

# ==================== AI ====================
google-genai==1.67.0          # Google Gemini API (latest)
anthropic>=0.87.0,<1.0        # Anthropic Claude API (CVE-2026-34450, CVE-2026-34452)
Pillow==12.1.1                # Image processing
beautifulsoup4==4.14.3        # HTML parsing for URL content (latest)
lxml==6.0.2                   # Fast HTML parser
numpy==2.4.2                  # Numerical operations (for RAG embeddings, latest)
faiss-cpu==1.13.2             # Vector similarity search (latest)
imageio[ffmpeg]==2.37.2       # GIF to video conversion for Gemini (latest)

# ==================== Music ====================
yt-dlp==2026.3.13             # YouTube downloader — IMPORTANT: old versions freeze on changed APIs
spotipy==2.25.2               # Spotify API wrapper (latest)
PyNaCl>=1.6.2,<2               # Voice encryption for discord.py voice

# ==================== Database ====================
aiosqlite==0.22.1             # Async SQLite (latest)

# ==================== Development ====================
watchdog==6.0.0               # File watcher for hot reload (latest)
colorama==0.4.6               # Windows ANSI color support (latest)
pyyaml>=6.0.1                 # YAML parsing for prompt templates
types-PyYAML>=6.0             # Type stubs for mypy
types-requests>=2.32.0.20250306  # Type stubs for requests/spotipy integration

# ==================== Monitoring (Optional) ====================
sentry-sdk==2.53.0            # Error tracking (optional)
prometheus-client==0.24.1     # Metrics & monitoring (latest)

# ==================== Performance ====================
orjson==3.11.7               # 10x faster JSON parsing (drop-in replacement)


# ==================== Windows Utilities ====================
winshell==0.6; sys_platform == 'win32'   # Windows shell integration
pywin32==311; sys_platform == 'win32'    # Windows COM automation

# ==================== Token Counting ====================
tiktoken==0.12.0              # Accurate token counting for AI (latest)

# ==================== Testing ====================
pytest==9.0.2                 # Test framework (latest)
pytest-asyncio==1.3.0         # Async test support (latest)
pytest-timeout==2.4.0         # Safety net: kills hung tests after 30s
pygments>=2.20.0              # Syntax highlighter — pinned for CVE-2026-4539

# ==================== Implicit Dependencies ====================
# These are auto-installed with the packages above:
# - aiosignal, frozenlist (aiohttp)
# - google-auth, googleapis-common-protos, protobuf (google-genai)
# - requests, urllib3, certifi (spotipy)
